Refactor AsyncExecutor to match GeventExecutor interface
Also refactored project layout a bit more.
